+++ backend/src/main/java/edu/bilkent/bilbilet/request/transaction/Refund.java
@@ -4,6 +4,8 @@ import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 
 @Data
 @AllArgsConstructor
@@ -11,8 +13,9 @@ public class Refund {
 
     @NotNull
     int travelerId;
+
     @NotNull
-    int amount;
+    BigDecimal amount;
 
 
 }
