+++ backend/src/main/java/edu/bilkent/bilbilet/request/transaction/Withdraw.java
@@ -4,6 +4,8 @@ import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 
 @Data
 @AllArgsConstructor
@@ -11,9 +13,10 @@ public class Withdraw {
 
     @NotNull
     int travelerId;
+
     @NotNull
     String ibanNumber;
 
     @NotNull
-    int amount;
+    BigDecimal amount;
 }
